move time-extraction test into its own file
